moved classes to structure code
