+++ src/services/cardsrecommendation/CostCalculatorService.js
@@ -1,4 +1,4 @@
-import CostCalculator from "../utils/CostCalculator.js";
+import CostCalculator from "../../utils/CostCalculator.js";
 
 class CostCalculationService {
     static calculateCosts(combinations) {
