--- src/utils/ShopPurchaseOptionsGenerator.js
@@ -1,6 +1,0 @@
-class ShopPurchaseOptionsGenerator {
-    static generateAllShopPurchaseOptions(...args) {
-        return args.reduce((a, b) => a.flatMap(d => b.map(e => [d, e].flat())));
-    }
-}
-export default ShopPurchaseOptionsGenerator;
