+++ tests/offlinehunter.js
@@ -1,14 +1,14 @@
-import CardDataService from "../src/services/CardDataService.js";
-import ShopPurchaseOptionsService from "../src/services/ShopPurchaseOptionsService.js";
-import ShopOutputService from "../src/services/ShopOutputService.js";
-import CostCalculationService from "../src/services/CostCalculatorService.js";
+import CardDataService from "../src/services/cardsrecommendation/CardDataService.js";
+import ShopPurchaseOptionsService from "../src/services/cardsrecommendation/ShopPurchaseOptionsService.js";
+import ShopOutputService from "../src/services/cardsrecommendation/ShopOutputService.js";
+import CostCalculationService from "../src/services/cardsrecommendation/CostCalculatorService.js";
 
 async function offlineHunt() {
     const cardData = CardDataService.readCardDataFromFile("./tests/fixtures/mock_data_ec_1.json");
 
     const productOptions = ShopPurchaseOptionsService.generateProductOptions(cardData);
 
-    const combinations = ShopPurchaseOptionsService.generateAllShopPurchaseOptions(productOptions);
+    const combinations = ShopPurchaseOptionsService.generateAllShopPurchaseOptions(...productOptions);
 
     const costs = CostCalculationService.calculateCosts(combinations);
